+++ tests/integration/test_media_repository_security.cpp
@@ -2,7 +2,11 @@
 
 #include "../support/registration_token.hpp"
 #include "merovingian/config/config.hpp"
-#include "merovingian/homeserver/vertical_slice.hpp"
+#include "merovingian/homeserver/auth_service.hpp"
+#include "merovingian/homeserver/local_http_router.hpp"
+#include "merovingian/homeserver/media_service.hpp"
+#include "merovingian/homeserver/room_service.hpp"
+#include "merovingian/homeserver/runtime.hpp"
 #include "merovingian/media/security.hpp"
 
 #include <catch2/catch_test_macros.hpp>